refactor: 버튼 코드 분리
